+++ streamdeck_ui/gui.py
@@ -24,7 +24,10 @@ try:
     from pynput.keyboard import Controller, Key
 except ImportError as pynput_error:
     pnput_supported = False
-    print("Virtual keyboard functionality disabled.")
+    print("---------------")
+    print("*** Warning ***")
+    print("---------------")
+    print("Virtual keyboard functionality has been disabled.")
     print("You can still run Stream Deck UI, however you will not be able to emulate key presses or text typing.")
     print("The most likely reason you are seeing this message is because you don't have an X server running")
     print("and your operating system uses Wayland.")