+++ streamdeck_ui/stream_deck_monitor.py
@@ -75,9 +75,35 @@ class StreamDeckMonitor:
 
     def _run(self):
         """Runs the internal monitor thread until completion"""
+        showed_open_help: bool = False
+        showed_enumeration_help: bool = False
+        showed_libusb_help: bool = False
         while not self.quit.is_set():
             with self.lock:
-                attached_streamdecks = DeviceManager.DeviceManager().enumerate()
+                attached_streamdecks = []
+                try:
+                    attached_streamdecks = DeviceManager.DeviceManager().enumerate()
+                    showed_libusb_help = False
+                except DeviceManager.ProbeError:
+                    if not showed_libusb_help:
+                        print("\n------------------------")
+                        print("*** Problem detected ***")
+                        print("------------------------")
+                        print("A suitable LibUSB installation could not be found.")
+                        print("Check installation instructions:")
+                        print("https://github.com/timothycrosley/streamdeck-ui")
+                        showed_libusb_help = True
+
+                        # No point showing the next help if we can't even enumerate
+                        showed_enumeration_help = True
+                        continue
+
+                if len(attached_streamdecks) == 0:
+                    if not showed_enumeration_help:
+                        print("No Stream Deck(s) detected. Attach a Stream Deck.")
+                        showed_enumeration_help = True
+                else:
+                    showed_enumeration_help = False
 
             # Look for new StreamDecks
             for streamdeck in attached_streamdecks:
@@ -86,8 +112,16 @@ class StreamDeckMonitor:
                     try:
                         self.attached(streamdeck_id, streamdeck)
                         self.streamdecks[streamdeck_id] = streamdeck
-                    except TransportError as error:
-                        print(f"Error during attach: {error}")
+                        showed_open_help = False
+                    except TransportError:
+                        if not showed_open_help:
+                            print("\n------------------------")
+                            print("*** Problem detected ***")
+                            print("------------------------")
+                            print("A Stream Deck is attached, but it could not be opened.")
+                            print("Check installation instructions and ensure a udev rule has been added and loaded.")
+                            print("https://github.com/timothycrosley/streamdeck-ui")
+                            showed_open_help = True
                         pass
 
             # Look for suspended/resumed StreamDecks