+++ CHANGELOG.md
@@ -23,7 +23,8 @@ Change log
 - Better support for dark/light mode (due to pyside6).
 ### Features
 - Added support for a new sub-variant of the StreamDeck XL.
-- Allow Stream Deck UI to start, even if virtual keyboard wont' work.
+- Allow Stream Deck UI to start, even if virtual keyboard won't work.
+- Improved troubleshooting messages.
 ## 2.0.6 - 23 Sep 2022
 ### Fixes
 - Image drag/drop from external applications.